[fix] responsiveness and routing issue
