--- frontend/app/MyNavBar/page.tsx
@@ -1,113 +1,0 @@
-"use client"
-import Link from '../../node_modules/next/link';
-import './../globals.css';
-import { Fragment, useState } from 'react'
-import { Dialog, Transition } from '@headlessui/react'
-import {
-  Bars3Icon,
-  Cog6ToothIcon,
-  DocumentDuplicateIcon,
-  FolderIcon,
-  XMarkIcon,
-} from '@heroicons/react/24/outline'
-
-const navigation = [
-  { name: 'Projects', href: '/Project', icon: FolderIcon, current: false },
-  { name: 'Inventory', href: '/Inventory', icon: DocumentDuplicateIcon, current: false },
-]
-
-function classNames(...classes: string[]): string {
-  return classes.filter(Boolean).join(' ');
-}
-
-export default function MyNavBar() {
-  const [sidebarOpen, setSidebarOpen] = useState(false)
-
-  return (
-    <div>
-      <Transition.Root show={sidebarOpen} as={Fragment}>
-        <Dialog as="div" className="relative z-50 lg:hidden" onClose={setSidebarOpen}>
-          <Transition.Child as={Fragment} enter="transition-opacity ease-linear duration-300" enterFrom="opacity-0" enterTo="opacity-100" leave="transition-opacity ease-linear duration-300" leaveFrom="opacity-100" leaveTo="opacity-0">
-            <div className="fixed inset-0 bg-gray-900/80" />
-          </Transition.Child>
-          <div className="fixed inset-0 flex">
-            <Transition.Child as={Fragment} enter="transition ease-in-out duration-300 transform" enterFrom="-translate-x-full" enterTo="translate-x-0" leave="transition ease-in-out duration-300 transform" leaveFrom="translate-x-0" leaveTo="-translate-x-full">
-              <Dialog.Panel className="relative mr-16 flex w-full max-w-xs flex-1">
-                <Transition.Child as={Fragment} enter="ease-in-out duration-300" enterFrom="opacity-0" enterTo="opacity-100" leave="ease-in-out duration-300" leaveFrom="opacity-100" leaveTo="opacity-0">
-                  <div className="absolute left-full top-0 flex w-16 justify-center pt-5">
-                    <button type="button" className="-m-2.5 p-2.5" onClick={() => setSidebarOpen(false)}>
-                      <span className="sr-only">Close sidebar</span>
-                      <XMarkIcon className="h-6 w-6 text-white" aria-hidden="true" />
-                    </button>
-                  </div>
-                </Transition.Child>
-                <div className="flex grow flex-col gap-y-5 overflow-y-auto bg-blue-950 px-6 pb-4">
-                  <div className="flex h-16 shrink-0 items-center">
-                    <Link href="/">
-                      <img className="h-60 mt-[20px] ml-[-5px] w-auto" src="image/epitech.png" alt="Your Company"/>
-                    </Link>
-                  </div>
-                  <nav className="flex flex-1 flex-col mt-[50px]">
-                    <ul role="list" className="flex flex-1 flex-col gap-y-7">
-                      <li>
-                        <ul role="list" className="-mx-2 space-y-1">
-                          {navigation.map((item) => (
-                            <li key={item.name}>
-                              <a href={item.href} className={classNames( item.current ? 'bg-indigo-700 text-white' : 'text-white hover:text-white hover:bg-indigo-700', 'group flex gap-x-3 rounded-md p-2 text-md leading-6 font-semibold mt-[20px]' )}>
-                                <item.icon className={classNames(   item.current ? 'text-white' : 'text-white group-hover:text-white',   'h-12 w-12 mt-[-10px] shrink-0' )} aria-hidden="true"/> {item.name}
-                              </a>
-                            </li>
-                          ))}
-                        </ul>
-                      </li>
-                      <li className="mt-auto">
-                        <a href="/Parameter" className="group -mx-2 flex gap-x-3 rounded-md p-2 text-md font-semibold leading-6 text-white hover:bg-indigo-700 hover:text-white">
-                          <Cog6ToothIcon className="h-12 w-12 mt-[-10px] shrink-0 text-white group-hover:text-white" aria-hidden="true" />
-                          Settings
-                        </a>
-                      </li>
-                    </ul>
-                  </nav>
-                </div>
-              </Dialog.Panel>
-            </Transition.Child>
-          </div>
-        </Dialog>
-      </Transition.Root>
-      <div className="hidden lg:fixed lg:inset-y-0 lg:z-50 lg:flex lg:w-72 lg:flex-col">
-        <div className="flex grow flex-col gap-y-5 overflow-y-auto bg-blue-950 px-6 pb-4">
-          <div className="flex h-16 shrink-0 items-center">
-            <Link href="/">
-              <img className="h-60 mt-[20px] ml-[-5px] w-auto" src="image/epitech.png" alt="Your Company" />
-            </Link>
-          </div>
-          <nav className="flex flex-1 flex-col mt-[50px]">
-            <ul role="list" className="flex flex-1 flex-col gap-y-7">
-              <li>
-                <ul role="list" className="-mx-2 space-y-1">
-                  {navigation.map((item) => (
-                    <li key={item.name}>
-                      <a href={item.href} className={classNames( item.current ? 'bg-indigo-700 text-white' : 'text-white hover:text-white hover:bg-indigo-700', 'group flex gap-x-3 rounded-md p-2 text-md leading-6 font-semibold mt-[20px]' )}>
-                        <item.icon className={classNames(   item.current ? 'text-white' : 'text-white group-hover:text-white',   'h-12 w-12 mt-[-10px] shrink-0' )} aria-hidden="true" /> {item.name}
-                      </a>
-                    </li>
-                  ))}
-                </ul>
-              </li>
-              <li className="mt-auto">
-                <a href="/Parameter" className="group -mx-2 flex gap-x-3 rounded-md p-2 text-md font-semibold leading-6 text-white hover:bg-indigo-700 hover:text-white" >
-                  <Cog6ToothIcon className="h-12 w-12 mt-[-10px] shrink-0 text-white group-hover:text-white" aria-hidden="true" />
-                  Settings
-                </a>
-              </li>
-            </ul>
-          </nav>
-        </div>
-      </div>
-        <button type="button" className="-m-2.5 p-2.5 text-gray-700 lg:hidden" onClick={() => setSidebarOpen(true)}>
-          <span className="sr-only">Open sidebar</span>
-          <Bars3Icon className="h-6 w-6" aria-hidden="true" />
-        </button>
-    </div>
-    );
-}
